extract into elf optimizer.
