+++ Cargo.toml
@@ -19,6 +19,7 @@ members = [
   "openvm-riscv/extensions/hints-transpiler",
   "openvm-riscv/extensions/hints-circuit",
   "trace-analyzer",
+  "elf-optimizer",
 ]
 
 exclude = ["riscv-runtime"]
@@ -43,6 +44,7 @@ powdr-autoprecompiles = { path = "./autoprecompiles", version = "0.1.4" }
 powdr-openvm-riscv = { path = "./openvm-riscv", version = "0.1.4" }
 powdr-openvm-bus-interaction-handler = { path = "./openvm-bus-interaction-handler", version = "0.1.4" }
 powdr-openvm = { path = "./openvm", version = "0.1.4" }
+powdr-elf-optimizer = { path = "./elf-optimizer", version = "0.1.4" }
 
 powdr-openvm-riscv-hints-guest = { path = "./openvm-riscv/extensions/hints-guest", version = "0.1.4" }
 powdr-openvm-riscv-hints-transpiler = { path = "./openvm-riscv/extensions/hints-transpiler", version = "0.1.4" }
